Add Landing repository (#11)
- add landing repository
- add book ui and mapper

+++ app/src/main/java/com/timetonic/test/landing/data/remote/model/BooksInfoDTO.kt
@@ -3,7 +3,7 @@ package com.timetonic.test.landing.data.remote.model
 import com.google.gson.annotations.SerializedName
 
 data class BooksInfoDTO(
-    @SerializedName("sstamp") val id: Int,
+    @SerializedName("sstamp") val id: Long,
     @SerializedName("fpForm") val details: BookDetailsDTO,
     @SerializedName("ownerPrefs") val ownerPrefs: OwnerPrefsDTO,
 )
